+++ backend/Models/Flight.cs
@@ -6,13 +6,13 @@ namespace backend.Models;
 
 public partial class Flight
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
 
     public string FlightCode { get; set; } = null!;
 
-    public int DeparturePort { get; set; }
+    public long DeparturePort { get; set; }
 
-    public int ArrivalPort { get; set; }
+    public long ArrivalPort { get; set; }
 
     public DateTime DepartureTime { get; set; }
 
@@ -29,9 +29,9 @@ public partial class Flight
 
     public int BusinessClassSeatsAvailable { get; set; }
 
-    public int FlightsAirlineId { get; set; }
+    public long FlightsAirlineId { get; set; }
 
-    public int FlightsAirplaneId { get; set; }
+    public long FlightsAirplaneId { get; set; }
 
     public string IdempotencyKey { get; set; }
     [JsonIgnore]
