+++ backend/Repositories/FlightRepository.cs
@@ -26,13 +26,13 @@ namespace backend.Repositories
             return flights;
         }
 
-        public async Task<Flight?> GetFlightById(int id)
+        public async Task<Flight?> GetFlightById(long id)
         {
             var flight = await _context.Flights.FindAsync(id);
             return flight;
         }
 
-        public async Task<Flight?> GetFlightWithRelationshipsById(int id)
+        public async Task<Flight?> GetFlightWithRelationshipsById(long id)
         {
             var flight = await _context.Flights
                 .Include(flight => flight.FlightsAirline)
@@ -108,7 +108,7 @@ namespace backend.Repositories
             return flight;
         }
 
-        public async Task<Flight> Delete(int id)
+        public async Task<Flight> Delete(long id)
         {
             var transaction = _context.Database.BeginTransaction();
 
@@ -168,7 +168,7 @@ namespace backend.Repositories
         }
 
 
-        public async Task<List<Flight>> GetFlightsByDepartureDestinationAndDepartureDate(int departureAirportId, int destinationAirportId, DateOnly departureDate)
+        public async Task<List<Flight>> GetFlightsByDepartureDestinationAndDepartureDate(long departureAirportId, long destinationAirportId, DateOnly departureDate)
         {
             var flights = await _context.Flights
                 .Where(flight =>
@@ -183,7 +183,7 @@ namespace backend.Repositories
             return flights;
         }
 
-        public async Task<List<Flight>> GetFlightsByAirplaneId(int airplaneId)
+        public async Task<List<Flight>> GetFlightsByAirplaneId(long airplaneId)
         {
             var flights = await _context.Flights
                 .Where(flight => flight.FlightsAirplaneId == airplaneId)
@@ -192,13 +192,13 @@ namespace backend.Repositories
             return flights;
         }
 
-        public async Task<FlightClass?> GetFlightClassById(int id)
+        public async Task<FlightClass?> GetFlightClassById(long id)
         {
             var flightClass = await _context.FlightClasses.FindAsync(id);
             return flightClass;
         }
 
-        public async Task<List<Ticket>> GetTicketsByFlightId(int flightId)
+        public async Task<List<Ticket>> GetTicketsByFlightId(long flightId)
         {
             var tickets = await _context.Tickets.Where(ticket => ticket.FlightId == flightId)
                 .Include(ticket => ticket.Passenger)
