+++ backend/Repositories/IAirplaneRepository.cs
@@ -7,6 +7,6 @@ namespace backend.Repositories
     {
         Task<List<Airplane>> GetAll();
 
-        Task<Airplane?> GetAirplaneById(int id);
+        Task<Airplane?> GetAirplaneById(long id);
     }
 }
